Move DgraphBundledException into dgraph_bundled.py
